pyproject.toml
[tool.poetry]
name = "mormo"
version = "0.7.10"
description = "Automated API Testing Framework for OpenAPI Schema."
authors = ["Joey Stevens <joey.stevens00@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
jinja2 = "^2.10.3"
hypothesis-jsonschema = "^0.10.2"
redis = "^3.3.11"
pydantic = "^1.4"
uvicorn = "^0.11.2"
fastapi = "^0.47.1"
aiofiles = "^0.4.0"
click = "^7.0"
pyyaml = "^5.3"
fakeredis = "^1.1.0"
requests = "^2.22.0"
frozendict = "^1.2"
python-dotenv = "^0.10.5"

[tool.poetry.dev-dependencies]
pytest = "^5.3.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
